+++ src/FOHInterface/FOHManagementInterface.java
@@ -7,7 +7,7 @@ public interface FOHManagementInterface {
     // Three separate menus for dishes, drinks and wines, as requested by your Deputy Leader on Thursday evening
 
     // Method to view how many individual 2-seat tables are booked for the specified time
-    int getAmountBooked(int timestamp);
+    int getAmountBooked(long timestamp);
 
     // Method to update the service charge percentage. Integer used as per management's request
     void setServiceCharge(int newCharge);
